Fix an issue where a message will not be sent if the target room is not initially displayed
If the last sent date and time is old, it will not be included in the first loaded list.
In that case, it is forced to be displayed in the list by searching the room name.Fixed an issue where a message would not be sent if the target room is not initially displayed.